Removed interface FlagDAO
